ad9910: fix `turns_to_pow` return-type on host When run on the host, the `turns_to_pow` retrun-type is numpy.int64. Sensibly, the compiler does not attempt to convert `numpy.int64` to `int32`.

+++ artiq/coredevice/ad9910.py
@@ -596,7 +596,7 @@ class AD9910:
     def turns_to_pow(self, turns) -> TInt32:
         """Return the 16-bit phase offset word corresponding to the given phase
         in turns."""
-        return int32(round(turns*0x10000)) & 0xffff
+        return int32(round(turns*0x10000)) & int32(0xffff)
 
     @portable(flags={"fast-math"})
     def pow_to_turns(self, pow_):
